+++ Frends.Excel.Parse/Frends.Excel.Parse/Excel.cs
@@ -1,6 +1,4 @@
 ﻿using System.ComponentModel;
-using System.Data;
-using System.Linq;
 using System.Text;
 using ExcelDataReader;
 using Frends.Excel.Parse.Definitions;
@@ -18,9 +16,8 @@ public static class Excel
     /// </summary>
     /// <param name="input">Input configuration</param>
     /// <param name="options">Input options</param>
-    /// <param name="cancellationToken"></param>
-    /// <returns>Result containing the parsed Excel workbook data.</returns>
-    /// <exception cref="Exception"></exception>
+    /// <param name="cancellationToken">Cancellation token to cancel the operation.</param>
+    /// <returns>Result containing the parsed Excel: object { bool Success, string ErrorMessage, DataSet DataSet }</returns>
     public static Result Parse(
         [PropertyTab] Input input,
         [PropertyTab] Options options,
@@ -34,36 +31,18 @@ public static class Excel
             cancellationToken.ThrowIfCancellationRequested();
             Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
 
-            using var stream = new FileStream(input.Path, FileMode.Open, FileAccess.Read);
-            using var excelReader = ExcelReaderFactory.CreateReader(stream);
-            var result = ConvertToWorkbookData(excelReader.AsDataSet(), cancellationToken);
-            return new Result(true, result);
+            using (var stream = new FileStream(input.Path, FileMode.Open, FileAccess.Read))
+            {
+                using (var excelReader = ExcelReaderFactory.CreateReader(stream))
+                {
+                    var result = excelReader.AsDataSet();
+                    return new Result(true, result, null);
+                }
+            }
         }
         catch (Exception ex)
         {
             return ex.Handle(options);
         }
-    }
-
-    private static WorkbookData ConvertToWorkbookData(DataSet dataSet, CancellationToken cancellationToken)
-    {
-        return new WorkbookData
-        {
-            Tables = dataSet.Tables
-                .Cast<DataTable>()
-                .Select(table =>
-                {
-                    cancellationToken.ThrowIfCancellationRequested();
-                    return new WorksheetData
-                    {
-                        TableName = table.TableName,
-                        Columns = table.Columns.Cast<DataColumn>().Select(column => column.ColumnName).ToList(),
-                        Rows = table.Rows.Cast<DataRow>()
-                            .Select(row => row.ItemArray.Select(value => value == DBNull.Value ? null : value).ToList())
-                            .ToList(),
-                    };
-                })
-                .ToList(),
-        };
     }
 }