+++ Frends.Excel.Parse/Frends.Excel.Parse/Helpers/ErrorHandler.cs
@@ -9,6 +9,9 @@ namespace Frends.Excel.Parse.Helpers;
 /// </summary>
 internal static class ErrorHandler
 {
+    /// <summary>
+    /// Converts an exception into a failed Result object or rethrows based on task options.
+    /// </summary>
     /// <param name="exception">The exception to handle.</param>
     /// <param name="options">Task options that control whether failures are returned as a Result object or thrown.</param>
     /// <param name="throwCanceled">
@@ -29,7 +32,7 @@ internal static class ErrorHandler
         if (throwCanceled && exception is OperationCanceledException) throw exception;
     }
 
-    private static void ThrowBaseException(Exception exception, string? customMessage = null)
+    private static void ThrowBaseException(Exception exception, string customMessage = null)
     {
         if (string.IsNullOrEmpty(customMessage))
             ExceptionDispatchInfo.Capture(exception).Throw();
@@ -37,7 +40,7 @@ internal static class ErrorHandler
         throw new Exception(customMessage, exception);
     }
 
-    private static Result ReturnResult(Exception exception, string? customMessage = null)
+    private static Result ReturnResult(Exception exception, string customMessage = null)
     {
         var errorMessage = string.IsNullOrEmpty(customMessage)
             ? exception.Message
@@ -45,6 +48,7 @@ internal static class ErrorHandler
 
         return new Result(
             false,
+            null,
             error: new Error
             {
                 Message = errorMessage,